Refactoring to adopt new wsman lib changes.
