+++ src/main/java/com/dell/isg/smi/adapter/server/config/ServerAdapterConstants.java
@@ -1,7 +1,7 @@
 /**
  * Copyright © 2017 DELL Inc. or its subsidiaries.  All Rights Reserved.
  */
-package com.dell.isg.smi.adapter.server;
+package com.dell.isg.smi.adapter.server.config;
 
 /**
  * @author rajesh.varada
